fraggpt2/requirements.txt
absl-py==1.4.0
accelerate==0.11.0
adapter-transformers==3.2.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
albumentations==0.5.2
altair==5.0.1
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
ase==3.22.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.79
bleach==6.0.0
braceexpand==0.1.7
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.15.0
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.9
einops==0.6.0
executing==1.2.0
fastapi==0.98.0
fastBPE==0.1.0
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.11.0
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.31
-e git+https://github.com/openai/glide-text2im.git@69b530740eb6cef69442d6180579ef5ba9ef063e#egg=glide_text2im
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googledrivedownloader==0.4
gradio==3.23.0
gradio_client==0.2.7
grpcio==1.56.0
h11==0.14.0
h5py==2.10.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
hydra-core==1.1.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.5.0
libclang==16.0.0
lightning-utilities==0.8.0
lightseq==3.0.1
linkify-it-py==2.0.2
littleutils==0.2.2
llvmlite==0.39.1
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdown2==2.4.9
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.2.0
mmtf-python==1.1.3
moleculekit==1.6.9
more-itertools==9.1.0
mosestokenizer==1.2.1
mpi4py==3.1.4
msgpack==1.0.5
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
nh3==0.2.13
ninja==1.11.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
ogb==1.3.6
omegaconf==2.1.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openfile==0.0.7
opt-einsum==3.3.0
orjson==3.9.1
outdated==0.2.2
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
-e git+https://github.com/bytedance/ParaGen.git@4e33a900bcad1a54657855ece3bc4342628c7e13#egg=paragen
parso==0.8.3
pathtools==0.1.2
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
portalocker==2.7.0
ProDy==2.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysmiles==1.0.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-louvain==0.16
python-multipart==0.0.6
pytorch-lightning==1.2.9
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rdflib==6.3.2
rdkit==2022.9.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rouge==1.0.1
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.9.3
seaborn==0.12.2
-e git+https://github.com/geekyutao/Inpaint-Anything.git@d95664e5792288ce10aacb337370856d51a33f6a#egg=segment_anything&subdirectory=segment_anything
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.98
sentry-sdk==1.19.1
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
toolwrapper==2.1.0
toolz==0.12.0
torch==1.12.1+cu113
torch-geometric==1.6.3
torch-scatter @ file:///mnt/d/pkg/torch_scatter-2.1.0%2Bpt112cu113-cp38-cp38-linux_x86_64.whl
torch-sparse @ file:///mnt/d/pkg/torch_sparse-0.6.15%2Bpt112cu113-cp38-cp38-linux_x86_64.whl
torchaudio==0.12.1+cu113
torchmetrics==0.2.0
torchvision==0.13.1+cu113
tornado==5.1.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
uctools==1.3.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
wandb==0.14.2
wavedrom==2.0.3.post3
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wldhx.yadisk-direct==0.0.6
wrapt==1.12.1
xtb==22.1
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.15.0
